Changed fetchedGame model and apiService
