+++ PriceTrackerMobile/Views/AddItemPage.xaml
@@ -46,10 +46,10 @@
                         </SwipeView.RightItems>
                         <StackLayout Orientation="Horizontal">
                             <Image
-                            Source="{Binding ImageUrl}"/>
+                                Source="{Binding ImageUrl}"/>
                             <Label
-                            Text="{Binding Name}"
-                            FontSize="Title" />
+                                Text="{Binding Name}"
+                                FontSize="Title" />
                         </StackLayout>
                     </SwipeView>
                 </DataTemplate>
